Let's shoot again

+++ src/Rasa.Game/Packets/MapChannel/Client/RequestPerformAbilityPacket.cs
@@ -9,7 +9,7 @@
 
         public ActionId ActionId { get; set; }
         public int ActionArgId { get; set; }
-        public long Target { get; set; }
+        public uint Target { get; set; }
         public int ItemId { get; set; }
 
         public override void Read(PythonReader pr)
@@ -17,7 +17,7 @@
             pr.ReadTuple();
             ActionId = (ActionId)pr.ReadInt();
             ActionArgId = pr.ReadInt();
-            Target = pr.ReadLong();
+            Target = (uint)pr.ReadLong();
             if (pr.PeekType() == PythonType.Int)
                 ItemId = pr.ReadInt();
             else
